FileAPI/url/url_createobjecturl_blob.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>FileAPI Test: url</title>
    <link rel="author" title="Intel" href="www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/FileAPI/#dfn-createObjectURL">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>

    <script>
      var blob = new Blob(["Test Blob"]);

      test(function() {
        var testBlob = window.URL.createObjectURL(blob);
        assert_equals(typeof testBlob, "string", "Blob URI is typeof string");
        assert_equals(testBlob.indexOf("blob"), 0, "Blob URI starts with 'blob'");
      }, "Check if the Blob URI starts with 'blob' without 'autoRevoke'");

      test(function() {
        var testBlob = window.URL.createFor(blob);
        assert_equals(typeof testBlob, "string", "Blob URI is typeof string");
        assert_equals(testBlob.indexOf("blob"), 0, "Blob URI starts with 'blob'");
      }, "Check if the Blob URI starts with 'blob' with autoRevoke is false");
    </script>
  </body>
</html>
